Move some code to tagfs_file_misc.c, to avoid gcc optimization problems
When I try -O0, copy_from_user() gets compile errors. So I moved
the code that needs that function to a separate file. But it turns
out the module still doesn't work (__init isn't called, and the
insmod wedges). Leaving this in a separate file for now. If this FS
survives, some refactoring decisions should be made in the future.