Switch to actually installing the kmod and using modprobe
I hope this won't cause new failures but it's a necessary risk. The
new scripts/install_kmod.sh script will install the kmod if the running
kernel does not have a native famfs module - if a native famfs is
present you can't use the out-of-tree version. This paves the way to get
famfs into a kernel for refining the patch set.