Cache coherency: flush/invalidate processor cache as necessary
fsck used read by default, which causes cache problems

Three cli commands have a --read|--mmap option: mount, logplay and fsck.
They consume the log and superblock either by read or mmap depending on
the option. Consuming by read is useful in tests, but it causes cache
coherency problems if the fs is in shared memory (because we can't
call clflush() on the source for the read, but we can call it on the
address from an mmap).

fsck actually defaulted to read, but now it defaults to mmap. All three
still support read, but they print a warning to stderr if you do this.

This commit also changes the behavior of 'famfs creat' as follows:
If you re-run 'famfs creat' on an existing file, with the -r and -S
options (randomize with a seed), it will re-write the data based on the
(presumably new) seed to the file. This is useful in cache coherency
testing.

Also updated the smoke tests to adapt to this change.

mkfs: kill_super needs a cache flush to make the killed superblock
visible