famfs: create, store and access system uuid information
To determine if the current system is a master or client, famfs cli
has to get system uuid from /sys/devices/virtual/dmi/id/product_uuid
which requires root privileges. To avoid this dependency generate a
new uuid and store it in a famfs specific location such that the
writer to famfs does not always need to be root. If the file is lost,
that famfs node will act as a client node.

Signed-off-by: Aravind Ramesh <arramesh@micron.com>