pcq: major smoke test update
Notably, after creating a set of pcq's, this runs pcq as non-root
for the rest of the tests. This required a "chown -R" before the
non-root runs to make coverage tests work - because some files
created by gcov are owned by root if the runs were as root.
This was a pain to figure out...