Use threadpool data movement in 'famfs cp'
This adds a threadpool to the locked_log structure, if thread_ct > 0.
Data movement is handled by the threadpool. When
famfs_release_locked_log() is called, the FLOCK on the log is released,
and thpool_wait(lp->thp) is called to wait for completion of all data
movement. 

This releases the log lock after all files have been created, but before
all data has been moved, shortening the interval during which the log is
locked.

Performance should improve due to multi-thread copy as well as shorter
log lock interval.