famfs mount: Make default fuse metadata timeout long
This more closely resembles the standalone famfs behavior, and will
likely put a bandaid on the "yaml leakage" bug - where the size of the
shadow yaml file leaks out and is seen instead of the correct size of a
file. This bug is almost certainly rooted in a race around timeing out
famfs_meta, and re-retrieving it via get_fmap.

Tests for that race, and a fix, are needed. The race (if it is indeed a
race) is in the kernel implementation.