Add support for attribute changes (chown, chmod etc.)
This is done by changing the attributes in the icache, and pinning
icache entries (files) that have had their attrs changed. This means
that once a file's attrs have been changed (chmod, chown, etc.) attrs
will be retrieved from the icache for the life of the mount (rather than
from the shadow yaml).

Other misc:
- Add FAMFS_ASSERT() macro that logs the issue before crashing famfs_fused
- Clean up some gratuitous logging
- icache: merge redundant putref functions

Fixes #104