Add dummy mount support and enable mkfs.famfs to use it
A dummy mount generates a tmp dir mount point, and does not check the
superblock or log. By doing this, mkfs.famfs and raw fsck can run
without raw daxdev access - they access the superblock and log via their
meta files.

mount --dummy does this (and outputs the mount point)
mkfs.famfs --nodax uses a dummy mount rather than raw dax access

This is added because the changes required to make fsdax work on devdax
in 6.15-onward *currently* break raw devdax mmap. That may be fixable,
and will be migrated into a new 'famfs' mode of devdax, but this was a
clean way not to stall while finishing up that work - and may be
necessary long term...